# expression: implement vectorized evaluation for builtinUTCTime*