<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for
`builtinUTCTimeWithArgSig`
`builtinUTCTimestampWithArgSig`
`builtinUTCTimeWithoutArgSig`
`builtinUTCTimestampWithoutArgSig`
for #12101

### What is changed and how it works?
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args 'builtinUTCTimeWithArgSig' 'builtinUTCTimestampWithArgSig' 'builtinUTCTimeWithoutArgSig' 'builtinUTCTimestampWithoutArgSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.0119 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimestampWithoutArgSig-VecBuiltinFunc-4         	  837919	      1243 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimestampWithoutArgSig-NonVecBuiltinFunc-4      	    3594	    327098 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimestampWithArgSig-VecBuiltinFunc-4            	    8982	    125110 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimestampWithArgSig-NonVecBuiltinFunc-4         	    3343	    357949 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimeWithoutArgSig-VecBuiltinFunc-4              	  338055	      3732 ns/op	     224 B/op	      11 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimeWithoutArgSig-NonVecBuiltinFunc-4           	     357	   3440515 ns/op	  229432 B/op	   11264 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimeWithArgSig-VecBuiltinFunc-4                 	     415	   2816045 ns/op	  196686 B/op	   10241 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinUTCTimeWithArgSig-NonVecBuiltinFunc-4              	     337	   3556657 ns/op	  229428 B/op	   11264 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	11.566s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test